+++ src/components/contact/ContactSelector.tsx
@@ -3,9 +3,8 @@ import React, { useState } from 'react';
 import { connect } from 'react-redux';
 import { Dropdown, DropdownProps } from 'semantic-ui-react';
 import { ContactSummary } from '../../clients/server.generated';
-import { formatContactName } from '../../helpers/contact';
+import { formatContactName, formatFunctionShort } from '../../helpers/contact';
 import { RootState } from '../../stores/store';
-import { getCompanyName } from '../../stores/company/selectors';
 
 interface Props {
   disabled?: boolean;
@@ -14,7 +13,6 @@ interface Props {
   placeholder: string;
   options: ContactSummary[];
   onChange: (value: number | number[]) => void;
-  getCompanyName: (id: number) => string;
 }
 
 function ContactSelector(props: Props & DropdownProps) {
@@ -28,7 +26,7 @@ function ContactSelector(props: Props & DropdownProps) {
     .map((x) => ({
       key: x.id,
       text: formatContactName(x.firstName, x.lastNamePreposition, x.lastName),
-      description: props.getCompanyName(x.companyId),
+      description: formatFunctionShort(x.function),
       value: x.id,
     }));
 
@@ -57,7 +55,6 @@ ContactSelector.defaultProps = {
 
 const mapStateToProps = (state: RootState) => ({
   options: state.summaries.Contacts.options,
-  getCompanyName: (id: number) => getCompanyName(state, id),
 });
 
 export default connect(mapStateToProps)(ContactSelector);